+++ editor/webapp/common/components/form-controls/input-text.vue
@@ -25,7 +25,11 @@
 import InputMixin from './inputMixin.js'
 
 export default {
-  name      : 'input-text',
+  name      : 'InputText',
+  components: {},
+  filters   : {},
+  mixins    : [InputMixin],
+  model     : {},
   props     : {
     value   : {
       type   : String,
@@ -60,18 +64,14 @@ export default {
   data      : function () {
     return {};
   },
-  model     : {},
+  computed  : {},
   created   : function () {
   },
-  computed  : {},
   methods   : {
     update(e) {
       this.$emit('input', e);
     }
-  },
-  components: {},
-  filters   : {},
-  mixins    : [InputMixin]
+  }
 }
 </script>